broke out control package.  :: sigh ::
decided I liked it after all when I broke out the layout implementations
into seperate classes.
