+++ explorer/src/main/java/com/chiralbehaviors/layout/explorer/AutoLayoutController.java
@@ -28,7 +28,7 @@ import javax.ws.rs.client.WebTarget;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.chiralbehaviors.layout.AutoLayout;
+import com.chiralbehaviors.layout.control.AutoLayout;
 import com.chiralbehaviors.layout.graphql.GraphQlUtil;
 import com.chiralbehaviors.layout.schema.Relation;
 import com.chiralbehaviors.layout.schema.SchemaNode;
