+++ kramer/src/main/java/com/chiralbehaviors/layout/control/OutlineSkin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.chiralbehaviors.layout;
+package com.chiralbehaviors.layout.control;
 
 import javafx.scene.control.SkinBase;
