+++ kramer/src/test/java/com/chiralbehaviors/layout/schema/Smoke.java
@@ -16,7 +16,7 @@
 
 package com.chiralbehaviors.layout.schema;
 
-import com.chiralbehaviors.layout.AutoLayout;
+import com.chiralbehaviors.layout.control.AutoLayout;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import javafx.application.Application;
