+++ toy-app/src/main/java/com/chiralbehaviors/layout/toy/SinglePageApp.java
@@ -29,8 +29,8 @@ import javax.ws.rs.client.WebTarget;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.chiralbehaviors.layout.AutoLayout;
 import com.chiralbehaviors.layout.Layout.LayoutModel;
+import com.chiralbehaviors.layout.control.AutoLayout;
 import com.chiralbehaviors.layout.graphql.GraphQlUtil.QueryException;
 import com.chiralbehaviors.layout.schema.Relation;
 import com.chiralbehaviors.layout.toy.Page.Route;
